+++ lib/matmath/src/main/java/matmath/MatMutator.java
@@ -2,6 +2,7 @@ package matmath;
 
 import org.opencv.core.Mat;
 
+@FunctionalInterface
 public interface MatMutator {
     void consume(int x, int y, Mat... m);
 }
